+++ uui-core/src/types/components/index.ts
@@ -1,4 +1,3 @@
-export * from './Button';
 export * from './Checkbox';
 export * from './TextInput';
 export * from './Spinner';