+++ uui/components/buttons/IconButton.tsx
@@ -1,4 +1,4 @@
-import { IconButtonBaseProps, IconButton as uuiIconButton } from '@epam/uui-components';
+import * as uuiComponents from '@epam/uui-components';
 import { withMods } from '@epam/uui-core';
 import css from './IconButton.module.scss';
 
@@ -11,7 +11,7 @@ interface IconButtonMods {
 }
 
 /** Represents the Core properties of the IconButton component. */
-export type IconButtonCoreProps = IconButtonBaseProps;
+export type IconButtonCoreProps = uuiComponents.IconButtonProps;
 
 /** Represents the properties of the IconButton component. */
 export type IconButtonProps = IconButtonCoreProps & IconButtonMods;
@@ -20,4 +20,4 @@ function applyIconButtonMods(mods: IconButtonProps & IconButtonMods) {
     return ['uui-icon_button', `uui-color-${mods.color || 'neutral'}`, css.root];
 }
 
-export const IconButton = withMods<IconButtonProps, IconButtonMods>(uuiIconButton, applyIconButtonMods);
+export const IconButton = withMods<IconButtonProps, IconButtonMods>(uuiComponents.IconButton, applyIconButtonMods);