+++ uui/components/buttons/__tests__/TabButton.test.tsx
@@ -1,16 +1,23 @@
 import React from 'react';
 import { TabButton } from '../TabButton';
-import { renderer } from '@epam/uui-test-utils';
+import { renderSnapshotWithContextAsync } from '@epam/uui-test-utils';
 import { ReactComponent as calendarIcon } from '../../../icons/calendar-18.svg';
 
 describe('TabButton', () => {
-    it('should be rendered correctly', () => {
-        const tree = renderer.create(<TabButton />).toJSON();
+    it('should be rendered correctly', async () => {
+        const tree = await renderSnapshotWithContextAsync(<TabButton />);
         expect(tree).toMatchSnapshot();
     });
 
-    it('should be rendered correctly with props', () => {
-        const tree = renderer.create(<TabButton onClick={ jest.fn } icon={ calendarIcon } isDisabled={ false } withNotify={ true } />).toJSON();
+    it('should be rendered correctly with props', async () => {
+        const tree = await renderSnapshotWithContextAsync(
+            <TabButton
+                onClick={ jest.fn }
+                icon={ calendarIcon }
+                isDisabled={ false }
+                withNotify={ true }
+            />,
+        );
         expect(tree).toMatchSnapshot();
     });
 });